% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_eratosthenes_f.R
\name{sieve_eratosthenes_f}
\alias{sieve_eratosthenes_f}
\title{Sieve of Eratosthenes}
\usage{
sieve_eratosthenes_f(n = 100)
}
\arguments{
\item{n}{integer A given limit}
}
\description{
Return all prime numbers up to any given limit.
Example of Fortran use
}
\examples{
sieve_eratosthenes_f(n = 100)

}
